split seal package into Schema and DSL
