+++ src/SEAL/Adapter/ReadWrite/ReadWriteConnection.php
@@ -4,8 +4,8 @@ namespace Schranz\Search\SEAL\Adapter\ReadWrite;
 
 use Schranz\Search\SEAL\Adapter\ConnectionInterface;
 use Schranz\Search\SEAL\Schema\Index;
-use Schranz\Search\SEAL\Search\Result;
-use Schranz\Search\SEAL\Search\Search;
+use Schranz\Search\SEAL\DSL\Result;
+use Schranz\Search\SEAL\DSL\Search;
 
 /**
  * @internal This class should never be needed to be instanced manually.
