+++ src/SEAL/DSL/README.md
@@ -1,0 +1,1 @@
+# Schranz Search DSL
