+++ src/SEAL/DSL/composer.json
@@ -1,0 +1,20 @@
+{
+    "name": "schranz-search/dsl",
+    "description": "DSL component for schranz-search/seal package.",
+    "type": "library",
+    "license": "MIT",
+    "autoload": {
+        "psr-4": {
+            "Schranz\\Search\\SEAL\\Search\\": ""
+        }
+    },
+    "authors": [
+        {
+            "name": "Alexander Schranz",
+            "email": "alexander@sulu.io"
+        }
+    ],
+    "require": {
+        "php": "^8.1"
+    }
+}
