+++ src/SEAL/Schema/.gitignore
@@ -1,0 +1,3 @@
+/vendor/
+/composer.phar
+/phpunit.xml
