+++ src/SEAL/composer.json
@@ -1,6 +1,6 @@
 {
     "name": "schranz-search/seal",
-    "description": "Search Engine Abstraction Layer",
+    "description": "DSL Engine Abstraction Layer",
     "type": "library",
     "license": "MIT",
     "autoload": {
@@ -9,7 +9,9 @@
         },
         "exclude-from-classmap": [
             "/Adapter/Multi/",
-            "/Adapter/ReadWrite/"
+            "/Adapter/ReadWrite/",
+            "/DSL/",
+            "/Schema/"
         ]
     },
     "authors": [
@@ -19,7 +21,9 @@
         }
     ],
     "require": {
-        "php": "^8.1"
+        "php": "^8.1",
+        "schranz-search/dsl": "^0.1@dev",
+        "schranz-search/schema": "^0.1@dev"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5"
